feat: introduce flake.nix leveraging pyproject.nix
introduce flake.nix configuration leveraging the already existing
information in pyproject.toml, by using pyproject.nix to parse it and
generate the correct derivation. this way, any updates to
pyproject.toml should automatically be reflected in the flake
configuration, with low additional maintaining overhead required.

the only time when this may come as a problem is due to dependabot
constantly upgrading the packages to versions that do not exist yet in
nixpkgs. it may be required to force dependabot to also update the
nixpkgs version declared in the flake for it to not also break (or at
least to upgrade less often)