chore: make `flake.nix` aware of dev dependencies, including ruff and pre-commit
by adding `dependency-groups`, `pyproject.nix` can now parse the
dependency groups and add these to the project, by passing `groups =
["dev"]` when calling the renderer.

special care was needed in order to expose `ruff` and `pre-commit` to
the toplevel of the environment. by default, it will treat them as
python dependencies and hide them in the python wrapper. by using
`groupBy`, we split the dependency list in "python" dependencies and
top level ones, such that they can be exposed.

sadly, these toplevel packages are not exposed in `python.pkgs` so we
must manually override them in a case by case basis.