feat: improve typing definitions, introduce mypy in CI (#338)
* feat: do not supress callback exceptions raised from calling callbacks

* fix: add explicit clauses for ConnectionClosedOK in send and heartbeat

* fix: change import order

* fix: apply ruff format

* feat: introduce flake.nix leveraging pyproject.nix

introduce flake.nix configuration leveraging the already existing
information in pyproject.toml, by using pyproject.nix to parse it and
generate the correct derivation. this way, any updates to
pyproject.toml should automatically be reflected in the flake
configuration, with low additional maintaining overhead required.

the only time when this may come as a problem is due to dependabot
constantly upgrading the packages to versions that do not exist yet in
nixpkgs. it may be required to force dependabot to also update the
nixpkgs version declared in the flake for it to not also break (or at
least to upgrade less often)

* chore: make `flake.nix` aware of dev dependencies, including ruff and pre-commit

by adding `dependency-groups`, `pyproject.nix` can now parse the
dependency groups and add these to the project, by passing `groups =
["dev"]` when calling the renderer.

special care was needed in order to expose `ruff` and `pre-commit` to
the toplevel of the environment. by default, it will treat them as
python dependencies and hide them in the python wrapper. by using
`groupBy`, we split the dependency list in "python" dependencies and
top level ones, such that they can be exposed.

sadly, these toplevel packages are not exposed in `python.pkgs` so we
must manually override them in a case by case basis.

* chore: refactor `dependencies-for`

* feat: add nix develop CI job

ensure that `nix develop` works on CI

* fix: change job name to be more compatible with the other actions

* feat: setup infra and run `python -m pytest` to run tests in nix setup action

do not rely on poetry to run tests, we'd like to not rely on it from
within the nix environment

* fix: do not rely on `npx` as its not on nix develop

use manual `supabase start` command instead

* fix: add `--command` typo

* chore: change name of nix develop command part

* feat: add coverage information and upload it to coveralls

* chore: run nix setup tests in both ubuntu and macos

* chore: undo macos latest as it apparently does not work

for some reason, `supabase start` does not seem to work on
`macos-latest` host in CI

* chore: switch from pylsp + mypy to pyright (basepyright)

it seems that pyright is better and faster than mypy in most if not
all cases, and the only reason I wasn't using it before is because it
was not working with eglot's default config.

* feat: add `basedpyright` to `pyproject.toml` instead of flake.nix only

this makes it available to all users, preparing for including it into
the standard CI later on

* fix: change back to python-lsp-server with pylsp-mypy

* chore: improve type definitions files to make mypy happy

there's still a lot of issues that I intend to improve upon. this is
just an initial set of changes to ensure that `mypy` doesn't
complain. we should further strengthen the mypy checking
rules (currently there are a lot of `Any`s in the code) as the code improves.

* chore: add mypy to CI as a step

* fix: run mypy through poetry

* fix: check for phx_ref_prev before calling del

* fix: use `python -m mypy` instead of `mypy` directly

* fix: run type check after `make run_tests` so that `poetry install` is ran

* fix: `StrEnum` does not exist in python 3.9

* fix: remove `phx_ref` from presence dict

* fix: add deprecation warning for calling send with dicts

* fix: fix config payload, improve more types

* fix: make typing definitions compatible with python 3.9

also enforce that mypy uses python 3.9 only rules.

* fix: import annotations for 3.9 to not complain about type error

* fix: move mypy to Makefile `run_tests`, run it before actual tests

* format: apply ruff reformating

* fix: try explicitly annotating `Callback` as a `TypeAlias`

* fix: change `dict` to `Dict` type due to 3.9

* fix: finally, import annotations to stop runtime from breaking

* format: apply ruff format one last time

* format: reorder import order in test_connection

* format: trim whitespace