feat: validate server message (#345)
* chore: introduce `ServerMessage` and `ClientMessage` types

`ServerMessage` represents all the different message types that a
server might send to a client
`ClientMessage` represents all the different message types that a
client might send to a server

these are there to validate the payload types, enforcing the
correctness of the payloads

* fix: improve messages schema

now all of messages should be valid

* feat: introduce `ServerMessage` parsing

now, instead of parsing and dispatching the messages inside the
`AsyncRealtimeChannel._trigger` function, first parse the message
using pydantic's `ServerMessageAdapter` type adapter, and then
dispatch the callbacks inside `_handle_message`, based on which
message was parsed.

this required the removal of the `_on` function, which was used to
both define callbacks to messages, but also to define callbacks to
event refs, which was very confusing. this behavior was used precisely
to define callbacks that are called on acks.

now, AsyncPush'es are saved in
AsyncRealtimeChannel.messages_waiting_for_ack, and later the callbacks
registered by `AsyncPush.receive` (now saved directly in the push
object) are called when a `phx_reply` type message is received.

some logic was introduce to avoid this message buffer from growing
forever indefinetly, by removing these messages from the dictionary
once the callback is called, or once the timeout is reached.